backend/requirements.txt
django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
psycopg2==2.9.1